refactor: move prototype utilities to contrib
