+++ compact_memory/__init__.py
@@ -22,8 +22,6 @@ __all__ = [
     "AgenticChunker",
     "DEFAULT_MEMORY_PATH",
     "LocalChatModel",
-    "render_five_w_template",
-    "MemoryCueRenderer",
     "ExperimentConfig",
     "run_experiment",
     "HistoryExperimentConfig",
@@ -37,7 +35,6 @@ __all__ = [
     "PromptBudget",
     "CompressionStrategy",
     "NoCompression",
-    "PrototypeSystemStrategy",
     "ValidationMetric",
     "run_params_trial",
     "StrategyConfig",
@@ -60,8 +57,6 @@ _lazy_map = {
     "AgenticChunker": "compact_memory.chunker",
     "DEFAULT_MEMORY_PATH": "compact_memory.config",
     "LocalChatModel": "compact_memory.local_llm",
-    "render_five_w_template": "compact_memory.prototype.canonical",
-    "MemoryCueRenderer": "compact_memory.prototype.memory_cues",
     "ExperimentConfig": "compact_memory.experiments.config",
     "run_experiment": "compact_memory.experiment_runner",
     "HistoryExperimentConfig": "compact_memory.history_experiment",
@@ -75,7 +70,6 @@ _lazy_map = {
     "PromptBudget": "compact_memory.prompt_budget",
     "CompressionStrategy": "compact_memory.compression",
     "NoCompression": "compact_memory.compression",
-    "PrototypeSystemStrategy": "compact_memory.prototype_system_strategy",
     "ValidationMetric": "compact_memory.validation.metrics_abc",
     "run_params_trial": "compact_memory.hpo",
     "StrategyConfig": "compact_memory.compression",
