--- compact_memory/prototype/__init__.py
@@ -1,10 +1,0 @@
-"""Utilities tied to the prototype-based memory system."""
-
-from .canonical import render_five_w_template
-from .memory_cues import MemoryCueRenderer
-
-
-__all__ = [
-    "render_five_w_template",
-    "MemoryCueRenderer",
-]
