+++ contrib/prototype_utils/__init__.py
@@ -1,0 +1,6 @@
+"""Utilities for the prototype-based memory system."""
+
+from .canonical import render_five_w_template
+from .memory_cues import MemoryCueRenderer
+
+__all__ = ["render_five_w_template", "MemoryCueRenderer"]
