+++ contrib/prototype_utils/memory_cues.py
