+++ contrib/strategies/__init__.py
@@ -1,3 +1,3 @@
 """Experimental compression strategies."""
 
-__all__ = ["chained", "rationale_episode"]
+__all__ = ["chained", "rationale_episode", "prototype_system"]
