+++ examples/prototype_evolution_demo.py
@@ -6,7 +6,7 @@ from pathlib import Path
 from tempfile import TemporaryDirectory
 
 from compact_memory.vector_store import InMemoryVectorStore
-from compact_memory.prototype_system_strategy import PrototypeSystemStrategy
+from contrib.strategies.prototype_system import PrototypeSystemStrategy
 from compact_memory.embedding_pipeline import MockEncoder
 import compact_memory.agent as agent_mod
