+++ tests/test_memory_cues.py
@@ -1,4 +1,4 @@
-from compact_memory.prototype.memory_cues import MemoryCueRenderer
+from contrib.prototype_utils.memory_cues import MemoryCueRenderer
 
 
 def test_memory_cue_renderer_basic():
@@ -6,4 +6,3 @@ def test_memory_cue_renderer_basic():
     cues = renderer.render(["Budget 2025 planning", "Alpha Beta Gamma"])
     assert "<MEM id=Budget_2025 />" in cues
     assert "<MEM id=Alpha_Beta />" in cues
-
